project/requirements.txt
alembic==1.8.0
async-generator==1.10
attrs==22.1.0
bidict==0.22.0
bigjson==1.0.9
black==22.12.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
convert-html-to-dash==0.1.2
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-bootstrap-templates==1.0.7
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==1.16.0
dominate==2.7.0
email-validator==1.3.0
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==4.11.4
importlib-resources==5.10.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.4
openpyxl==3.0.10
outcome==1.2.0
pandas==1.5.1
pathspec==0.10.3
platformdirs==2.6.2
plotly==5.11.0
psycopg2-binary==2.9.3
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.2
python-socketio==5.6.0
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
scipy==1.10.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
suds==1.1.2
tenacity==8.1.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.1.2
wsproto==1.2.0
WTForms==3.0.1
xmltodict==0.13.0
zeep==4.2.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
